refactor: manage file structure and split up styles
