+++ react-app/package.json
@@ -36,6 +36,7 @@
     "husky": "^8.0.0",
     "lint-staged": "^15.0.2",
     "prettier": "3.0.3",
+    "sass": "^1.69.5",
     "typescript": "^5.0.2",
     "vite": "^4.4.5"
   }
