+++ react-app/src/App.tsx
@@ -1,6 +1,6 @@
 import "./App.css";
 import ErrorBoundary from "./components/ErrorBoundary/ErrorBoundary";
-import PokemonsSearch from "./components/PokemonSearch/pokemonsSearch";
+import PokemonsSearch from "./components/PokemonSearch/PokemonsSearch";
 
 export default function App() {
   return (
