+++ react-app/src/components/PokemonSearch/PokemonCard/PokemonCard.scss
@@ -1,0 +1,12 @@
+.pokemon {
+  width: 320px;
+  min-height: 270px;
+  background-color: rgb(184, 218, 218);
+  border-radius: 18px;
+}
+
+.pokemon__title,
+.pokemon__text {
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  margin: 10px 0;
+}
