+++ react-app/src/components/PokemonSearch/SearchBar/searchBar.scss
@@ -1,0 +1,12 @@
+.search-bar {
+  display: flex;
+  gap: 10px;
+  justify-content: center;
+  margin-bottom: 30px;
+}
+
+.search__input {
+  border-radius: 8px;
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  font-size: 18px;
+}
