+++ react-app/src/components/PokemonSearch/SearchResults/searchResults.scss
@@ -1,0 +1,13 @@
+.search-results {
+  display: flex;
+  justify-content: center;
+  gap: 15px;
+  margin: 20px 0;
+}
+
+.search-results__container {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 15px;
+}
